Add graticule, change resizing behaviour, refactor
Initialize and draw graticule (meridians and parallels). Uses default graticule (10 degrees diff).

Define map style attributes as Globe static properties and repalce

Give map container div a constant width and height instead of 100vw and 100vh. Before, resizing the window
would scale the container but not the svg inside. Also we want the map's width and height to be specified.
Maybe in the future, have a fullscreen globe version that responds to window resize events and a separate
constant width/height globe that can be embedded somewhere on a page.