Add point and highlight mode to modifiableglobe, support different point types
Highlight mode changes the colors of the countries that on hover. Add setHighlightMode(),
unsetHighlightMode(), and css hover rule.

Point mode adds a point to the globe on click. Add setPointMode(), and unsetPointMode(),
modify code that add/remove points to only do so when in point mode.

Support different point types. 3 different point types:
- ALWAYS_VISIBLE: points visible no matter where on globe they are (will appear even if on
  other side of globe beyond horizon)
- INVISIBLE_BEYOND_HORIZON: points disappear when they go beyond the horizon
- VISIBLE_ON_HORIZON: points will stay visible, but when they are beyond the horizon the
  appear on the horizon instead