Add EndOverlay
Add EndOverlay class. Shows the results of the game with  a summary of the questions asked,
scores, and the final score at the top. Positioned to the right of the globe. Also has buttons for
playing again and main menu.

Modify startRound() to push every question result to a results array which is used by the endOverlay.
Modify endGame() to show EndOverlay and set its contents and position.

Add EndOverlay styles to test.html.

Add globe.getMapBBox(). Used in EndOverlay.setPosition() to figure out where to position it.