Add GeoData.isSameCountry() and GeoData.getCountryId()
Checks if two countries should be considered the same. Uses GeoData.getCountryId()
and compares the ids of both countries.

Also add GeoData.getCountryId(). Gets the id of the country. Uses ADM0_A3 field as the id
for most countries. For dependencies, uses SVO_A3 field (which should be the ADM0_A3 field
of the sovereign country, the dependency belongs to).

Also change Question.computeScore() to use GeoData.isSameCountry(). So now if the question
is France, and the guess is French Polynesia -> should be correct now.